Sniff`: move `isNumericCalculation()` method to `TokenGroup`
Along the same lines as for the `is*Number()` functions, the `isNumericCalculation()` method is only needed by a few sniffs and the method checks tokens between a start and end point, so moving the method out of the `Sniff` class and into the `TokenGroup` class is appropriate.

This is a preparation step to be able to switch to using the upcoming PHPCSUtils sniff abstract classes in the future.

Includes changing all relevant function calls to the method and moving (and adjusting) the unit tests.
