+++ PHPCompatibility/Util/Tests/Helpers/TokenGroup/IsNumericCalculationUnitTest.inc
