+++ PHPCompatibility/Util/Tests/TestHelperPHPCompatibility.php
@@ -46,24 +46,4 @@ class TestHelperPHPCompatibility extends Sniff
     public function process(File $phpcsFile, $stackPtr)
     {
     }
-
-    /**
-     * Wrapper to make the protected parent::isNumericCalculation() method testable.
-     *
-     * @since 9.0.0
-     *
-     * @param \PHP_CodeSniffer\Files\File $phpcsFile The file being scanned.
-     * @param int                         $start     Start of the snippet (inclusive), i.e. this
-     *                                               token will be examined as part of the
-     *                                               snippet.
-     * @param int                         $end       End of the snippet (inclusive), i.e. this
-     *                                               token will be examined as part of the
-     *                                               snippet.
-     *
-     * @return bool
-     */
-    public function isNumericCalculation(File $phpcsFile, $start, $end)
-    {
-        return parent::isNumericCalculation($phpcsFile, $start, $end);
-    }
 }
